Update to version 3.3.1

+++ WICED/platform/ARM_CM3/CMSIS/core_cmInstr.h
@@ -445,7 +445,7 @@ __attribute__( ( always_inline ) ) __STATIC_INLINE int32_t __REVSH(int32_t value
   uint32_t result;
 
   __ASM volatile ("revsh %0, %1" : __CMSIS_GCC_OUT_REG (result) : __CMSIS_GCC_USE_REG (value) );
-  return(result);
+  return( (int32_t) result);
 #endif
 }
